# api/swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "version" : "0.4.0",
    "title" : "Trackless API",
    "contact" : {
      "name" : "wjtje",
      "url" : "https://www.wjtje.ga"
    },
    "license" : {
      "name" : "Apache 2.0 License",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths" : {
    "/access" : {
      "get" : {
        "tags" : [ "access" ],
        "summary" : "See where you have access to",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Access"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "access" ],
        "summary" : "Give some group access to a command",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "groupId" : {
                    "type" : "integer"
                  },
                  "access" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "accessId" : {
                      "type" : "integer"
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "something wrong with your request"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      }
    },
    "/access/{accessId}" : {
      "get" : {
        "tags" : [ "access" ],
        "summary" : "Display an access rule",
        "parameters" : [ {
          "in" : "path",
          "name" : "accessId",
          "description" : "ID of the access to return",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Access"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "access" ],
        "summary" : "Remove an access rule",
        "parameters" : [ {
          "in" : "path",
          "name" : "accessId",
          "description" : "ID of the access to return",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "404" : {
            "description" : "not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      }
    },
    "/access/group/{groupId}" : {
      "get" : {
        "tags" : [ "access" ],
        "summary" : "See where a group has access to",
        "parameters" : [ {
          "in" : "path",
          "name" : "groupId",
          "description" : "ID of the group to return",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Access"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      }
    },
...
    "/group" : {
      "get" : {
        "tags" : [ "group" ],
        "summary" : "Get all the groups on the server",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Group"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "group" ],
        "summary" : "Create a new group on the server",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "groupName" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created the new group",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "groupId" : {
                      "type" : "integer"
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Something is wrong with your input"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      }
    },
    "/group/{groupId}" : {
      "get" : {
        "tags" : [ "group" ],
        "summary" : "Get infomation about a single group",
        "parameters" : [ {
          "in" : "path",
          "name" : "groupId",
          "description" : "ID of the group to return",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Group"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid ID"
          },
          "404" : {
            "description" : "Group not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "group" ],
        "summary" : "Removes a single group from the server and moves all users to the default group",
        "parameters" : [ {
          "in" : "path",
          "name" : "groupId",
          "description" : "ID of the group to remove",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "400" : {
            "description" : "Invalid ID"
          },
          "404" : {
            "description" : "Group not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      },
      "patch" : {
        "tags" : [ "group" ],
        "summary" : "Edits a group name",
        "parameters" : [ {
          "in" : "path",
          "name" : "groupId",
          "description" : "ID of the group to edit",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "groupName" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "400" : {
            "description" : "Invalid ID"
          },
          "404" : {
            "description" : "Group not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      }
    },
    "/group/{groupId}/add/{userId}" : {
      "post" : {
        "tags" : [ "group" ],
        "summary" : "Add a user to a group",
        "parameters" : [ {
          "in" : "path",
          "name" : "groupId",
          "description" : "ID of the group ",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "in" : "path",
          "name" : "userId",
          "description" : "ID of the user",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "400" : {
            "description" : "Invalid ID"
          },
          "404" : {
            "description" : "Group or user not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Access" : {
        "type" : "object",
        "properties" : {
          "accessId" : {
            "type" : "integer"
          },
          "access" : {
            "type" : "string"
          }
        }
      },
...
      "Group" : {
        "type" : "object",
        "properties" : {
          "groupId" : {
            "type" : "integer"
          },
          "groupName" : {
            "type" : "string"
          },
          "users" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/User"
            }
          }
        }
      },
...
      "User" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "integer"
          },
          "firstname" : {
            "type" : "string"
          },
          "lastname" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "groupId" : {
            "type" : "integer"
          },
          "groupName" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearer_token" : {
        "type" : "apiKey",
        "name" : "api_key",
        "in" : "header"
      }
    }
  },
  "servers" : [ {
    "description" : "SwaggerHub API Auto Mocking",
    "url" : "https://virtserver.swaggerhub.com/wjtje/Trackless/0.4.0"
  } ]
}